requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
PyPDF2
google-generativeai
python-multipart
pydantic
sqlalchemy
streamlit
requests
python-dotenv
